% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{init_renv}
\alias{init_renv}
\title{Create an renv lockfile}
\usage{
init_renv(
  project = NULL,
  lockfile = renv::paths$lockfile(project = project),
  ...
)
}
\arguments{
\item{project}{The project directory. If NULL, then the active project
will be used. If no project is currently active, then the current working
directory is used instead.}

\item{lockfile}{The location where the generated lockfile should be written.
By default, the lockfile is written to a file called renv.lock in the
project directory. When NULL, the lockfile (as an R object) is returned
directly instead.}

\item{...}{All additional arguments are passed directly to \link[renv:snapshot]{renv::snapshot}.
Please see the documentation for that function for all relevant details.}
}
\value{
A string giving the location of the generated lockfile.
}
\description{
This function is nearly identical to \link[renv:snapshot]{renv::snapshot}, with a couple
tiny tweaks. All arguments are passed directly to \link[renv:snapshot]{renv::snapshot},
so the functionality is identical.
}
\examples{
\dontrun{
init_renv()
}
}
\seealso{
\link[renv:snapshot]{renv::snapshot} which this function relies on.
}
